% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueread.R
\name{calculate_residuals}
\alias{calculate_residuals}
\title{Title: Calculate residuals of the user rating - average Goodreads rating}
\usage{
calculate_residuals(data_frame)
}
\arguments{
\item{data_frame}{cleaned data frame (GR_clean).}
}
\value{
a data frame with residuals.
}
\description{
Description: Calculate the residuals of average Goodreads rating - user rating.
}
