% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueread.R
\name{clean_GR_ratings}
\alias{clean_GR_ratings}
\title{Title: Remove ratings of 0 from GR_data}
\usage{
clean_GR_ratings(data_frame)
}
\arguments{
\item{data_frame}{data frame with relevant Goodreads data.}
}
\value{
A data frame with ratings of 0 removed.
}
\description{
Description: If a book has a rating of 0, it was not rated yet because Goodreads
only allows users to give the lowest rating of 1, so these books should be removed
}
