% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueread.R
\name{dist_hist}
\alias{dist_hist}
\title{Title: Distribution histogram}
\usage{
dist_hist(data_frame)
}
\arguments{
\item{data_frame}{data frame with residuals data.}
}
\value{
A distribution histogram.
}
\description{
Description: Visually displays distribution histogram to see if user tends to be
more positive or negative than average Goodreads users.
}
