% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueread.R
\name{read_GR_data}
\alias{read_GR_data}
\title{Title: Read in and clean GoodReads data from CSV file}
\usage{
read_GR_data(file_path)
}
\arguments{
\item{file_path}{A character string representing the path to the CSV file.}
}
\value{
A data frame containing the cleaned Goodreads data.
}
\description{
Description: This function reads in data from user GoodReads CSV file
into a data frame and removes unnecessary columns.
}
