% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueread.R
\name{res_data}
\alias{res_data}
\title{Title: Calculate average, standard deviation, highest, and lowest residuals}
\usage{
res_data(data_frame)
}
\arguments{
\item{data_frame}{data frame with residuals data.}
}
\value{
average, highest, and lowest residual data and titles associated with data.
}
\description{
Description: Using the residuals data in the data frame,
calculate the average residual score, as well as the highest
and lowest scores.
}
