% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueread.R
\name{resids}
\alias{resids}
\title{Title: Calculate residuals of the average Goodreads rating - user rating}
\usage{
resids(data_frame)
}
\arguments{
\item{data_frame}{cleaned data frame (GR_clean).}
}
\value{
a data frame with residuals.
}
\description{
Description: Calculate the residuals of average Goodreads rating - user rating.
}
