% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueread.R
\name{read_GR_data}
\alias{read_GR_data}
\title{Title: Read in GoodReads data from CSV file
Description: This function reads in data from user GoodReads CSV file
into a data frame and removes unnecessary columns.}
\usage{
read_GR_data(file_name)
}
\arguments{
\item{file_name}{A character string representing the file path to the Goodreads CSV file.}
}
\value{
A data frame containing the cleaned Goodreads data.
}
\description{
Title: Read in GoodReads data from CSV file
Description: This function reads in data from user GoodReads CSV file
into a data frame and removes unnecessary columns.
}
